#8 Frontend Leaderboard - Refactor and updates for recent results display logic
