+++ game-engine/src/wordle/dtos.py
@@ -120,7 +120,7 @@ class LeaderboardEntryDto(BaseModel):
     total_golf_score: int = Field(description="Total golf score (lower is better)")
     first_game_date: date = Field(description="Date of first game")
     last_game_date: date = Field(description="Date of most recent game")
-    recent_form: list[RecentGameDto] = Field(default_factory=list, description="Last 5 games for recent form")
+    recent_results: list[RecentGameDto] = Field(default_factory=list, description="Last 5 games for recent results")
 
     model_config = ConfigDict(from_attributes=True)
